+++ biz/webui/htdocs/src/js/json-viewer.js
@@ -7,6 +7,7 @@ var CopyBtn = require('./copy-btn');
 var message = require('./message');
 var ContextMenu = require('./context-menu');
 var win = require('./win');
+var Tips = require('./panel-tips');
 
 var JSONTree = require('./components/react-json-tree')['default'];
 var dataCenter = require('./data-center');
@@ -16,6 +17,7 @@ var MAX_LENGTH = 1024 * 16;
 var STR_SELECTOR = 'span[style="color: rgb(133, 153, 0);"]';
 var LINK_RE = /^"(https?:)?(\/\/[^/]\S+)"$/i;
 var contextMenuList = [
+  { name: 'Copy Object' },
   { name: 'Expand All' },
   { name: 'Collapse All' }
 ];
@@ -49,7 +51,9 @@ var JsonViewer = React.createClass({
   },
   onContextMenu: function(e) {
     var isDialog = this.props.dialog;
-    var ctxMenu = util.getMenuPosition(e, 110, isDialog ? 60 : 90);
+    var data = this.props.data || {};
+    var ctxMenu = util.getMenuPosition(e, 110, isDialog ? 90 : 120);
+    contextMenuList[0].copyText = data.str || '';
     ctxMenu.list = isDialog ? contextMenuList : contextMenuList.concat(SEARCH_MENU);
     this.refs.contextMenu.show(ctxMenu);
     e.preventDefault();
@@ -221,9 +225,14 @@ var JsonViewer = React.createClass({
     var viewSource = state.viewSource;
     var props = this.props;
     var data = props.data;
+    var tips = props.tips;
+    var className = 'fill orient-vertical-box w-properties-wrap w-json-viewer';
     var noData = !data;
     if (noData) {
       data = state.lastData || {};
+      if (tips) {
+        return <div className={className + (props.hide ? ' hide' : '')}><Tips data={tips} /></div>;
+      }
     } else {
       state.lastData = data;
     }
@@ -240,11 +249,9 @@ var JsonViewer = React.createClass({
     }
     return (
       <div
-        className={
-          'fill orient-vertical-box w-properties-wrap w-json-viewer' +
-          (noData || props.hide ? ' hide' : '')
-        }
+        className={className + (noData || props.hide ? ' hide' : '')}
       >
+        <Tips data={tips} />
         <div className="w-textarea-bar">
           <CopyBtn value={data.str} />
           <a