+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# v2.9.97
+1. feat: 导入 Rules 和 Values 的交互调整，允许用户手动选择需要导入的内容
+2. feat: 安装插件功能优化
+3. feat: 支持通过 `lineProps://internal` 将规则作用于 Whistle 内部请求（只支持 host、proxy 等部分规则）
+
+# v2.9.96
+1. feat: 支持导出指定规则项
+
+# v2.9.95
+1. feat: JSON Viewer 的右键菜单支持 Copy 选中的文本
+2. feat: Composer 历史列表添加右键菜单
+3. feat: 优化界面及性能
+4. feat: 支持通过插件安装插件，实现方式参考：https://github.com/whistle-plugins/whistle.installer
+5. feat: 支持通过界面卸载插件
+
 # v2.9.94
 1. fix: https://github.com/avwo/whistle/pull/1192